Correcting errors, preparing for protection
