+++ js/main.js
@@ -2,6 +2,8 @@ import './miniature-picture.js';
 import './big-picture.js';
 import './api.js';
 import './upload.js';
+import './popup.js';
+import './filter.js';
 import {createSimilarFotos} from './miniature-picture.js';
 import {setUserFormSubmit} from './form.js';
 import {getCloseUploadCancel} from './popup.js';
