+++ js/upload.js
@@ -6,8 +6,8 @@ const imgUploadInput = document.querySelector('.img-upload__start input[type=fil
 imgUploadInput.addEventListener('change', () => {
   const file = imgUploadInput.files[0];
   const fileName = file.name.toLowerCase();
-
   const matches = FILE_TYPES.some((it) => fileName.endsWith(it));
+
   if (matches) {
     const reader = new FileReader();
     reader.addEventListener('load', () => {
@@ -15,5 +15,6 @@ imgUploadInput.addEventListener('change', () => {
     });
     reader.readAsDataURL(file);
   }
+
 });
